+++ python-package/lets_plot/frontend_context/_jupyter_notebook_ctx.py
@@ -11,23 +11,17 @@ from typing import Dict
 # noinspection PyPackageRequirements
 from IPython.display import display_html
 
-from .frontend_context import FrontendContext
-from .._global_settings import _get_global_str, _has_global_value, _is_production
+from ._frontend_ctx import FrontendContext
 from .. import _kbridge as kbr
+from .._global_settings import get_global_str, has_global_value, is_production
 from .._version import __version__
 
 
 class JupyterNotebookContext(FrontendContext):
 
-    def __init__(self, connected: bool) -> None:
+    def __init__(self, offline: bool) -> None:
         super().__init__()
-        self.connected = connected
-
-    def as_str(self, plot_spec: Dict) -> str:
-        return kbr._generate_dynamic_display_html(plot_spec)
-
-    def _undef_modules_script(self) -> str:
-        pass
+        self.connected = not offline
 
     def configure(self, verbose: bool):
         if self.connected:
@@ -37,13 +31,16 @@ class JupyterNotebookContext(FrontendContext):
             # noinspection PyTypeChecker
             display_html(self._configure_embedded_script(verbose), raw=True)
 
+    def as_str(self, plot_spec: Dict) -> str:
+        return kbr._generate_dynamic_display_html(plot_spec)
+
     @staticmethod
     def _configure_connected_script(verbose: bool) -> str:
-        base_url = _get_global_str("js_base_url")
-        if _has_global_value('js_name'):
-            name = _get_global_str('js_name')
+        base_url = get_global_str("js_base_url")
+        if has_global_value('js_name'):
+            name = get_global_str('js_name')
         else:
-            suffix = ".min.js" if _is_production() else ".js"
+            suffix = ".min.js" if is_production() else ".js"
             name = "lets-plot-{version}{suffix}".format(version=__version__, suffix=suffix)
 
         url = "{base_url}/{name}".format(base_url=base_url, name=name)