+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Windows platform support.
  - factor() function.
  - tooltip for geom_rect
+ 
+ ### Fixed
+ - error if data contains NaT value.
 
 ## [1.2.1] - 2020-02-13
 ### Fixed